+++ glam/urls.py
@@ -20,6 +20,7 @@ urlpatterns = [
     path("api/v1/data/", api_views.aggregations, name="v1-data"),
     path("api/v1/probes/", api_views.probes, name="v1-probes"),
     path("api/v1/probes/random/", api_views.random_probes, name="v1-random-probes"),
+    path("api/v1/updates/", api_views.updates, name="v1-updates"),
     path(
         "api/v1/front-end-auth-config/",
         auth_views.front_end_auth_config,